create all CRUD operations for Product model

+++ broker/models.py
@@ -23,7 +23,7 @@ class Salesman(models.Model):
 class Product(models.Model):
     name = models.CharField(max_length=50)
     description = models.TextField()
-    price = models.DecimalField(max_digits=4,decimal_places=2)
+    price = models.IntegerField()
     def __str__(self):
         return self.name
